18-Triplet-Link/pyproject.toml
[project]
name = "kb-solokb-solo"
version = "0.1.0"
description = "Knowledge Graph Construction Pipeline"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "pandas>=2.0.0",
    "requests>=2.31.0",
    "openai>=1.0.0",
    "pydantic>=2.0.0",
    "tenacity>=8.2.0",
    "python-dotenv>=1.0.0",
    "tqdm>=4.66.0",
    "langchain>=1.2.3",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]
